Move main logic to extension file and export required interfaces
